Created abstraction for AbstractProjectPullResource
