+++ src/main/resources/i18n/devhub_en.properties
@@ -54,6 +54,7 @@ section.commits = Commits
 section.pull-requests = Pull requests
 section.assignments = Assignments
 section.contributors = Contributors
+section.issues = Issues
 
 error.could-not-connect-git-server = Could not connect to the Git server!
 error.could-not-comment = Comment could not be persisted
